% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Data preprocessing}
\usage{
preprocess(data, factor = NULL, numeric = NULL, character = NULL,
  omit.missing = FALSE)
}
\arguments{
\item{data}{[\code{character(1)}]\cr
Give the data set to be preprocessing.}

\item{factor}{[\code{factor(1)}]\cr
A vector of variable names which should be as factor.
Default is \code{NULL}.}

\item{numeric}{[\code{factor(1)}]\cr
A vector of variable names which should be as numeric.
Default is \code{NULL}.}

\item{character}{[\code{factor(1)}]\cr
A vector of variable names which should be as character.
Default is \code{NULL}.}

\item{omit.missing}{[\code{logical(1)}]\cr
Omit the missings values?
Default is \code{FALSE}.}
}
\description{
Data preprocessing for factor, numeric, character and missing values.
}

